requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
SQLAlchemy==1.4.50
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
WTForms==3.0.1
PyJWT==2.8.0
Pillow==10.2.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
